2 bugs fixed: session keeper + deleting only own tickets.

+++ app/assets/javascripts/views/TicketSmall.js
@@ -5,7 +5,19 @@ MTS.Views.TicketViewSmall = Backbone.View.extend({
 		"dblclick .ticket_small":"deleteTicket"
 	},
 	deleteTicket: function(){
-		this.model.destroy({wait: true});
+		var user = window.sessionStorage.getItem("id"),
+		ticket_owner = this.model.get("user_id");
+		if (user)
+		{
+			if (parseInt(user) === ticket_owner)
+				this.model.destroy({wait: true});
+			else
+				console.log("It's not your ticket!");
+		}
+		else
+		{
+			console.log("Not authorized!!!");
+		}
 	},
 	render: function(container){
 		this.$el.html(this.template(this.model.toJSON()));
